+++ Sources/Molusca/Argument/Wizard.swift
@@ -1,6 +1,7 @@
 import ArgumentParser
 import ConsoleIO
 import Explorer
+import Mantle
 import XcodeTemplate
 
 struct Wizard: ParsableCommand {
