--- Sources/Molusca/Template+Extension.swift
@@ -1,7 +1,0 @@
-import XcodeTemplate
-
-extension Template {
-    func toUserDescription() -> String {
-        return "\(self.rawValue)). \(self.name)"
-    }
-}
